torque angle calculation working

+++ SimpleTime/lib/app/types/types.h
@@ -3,7 +3,7 @@
 
 namespace app{
     typedef unsigned long timestamp;
-    typedef unsigned int TorqueAngle;
+    typedef double TorqueAngle;
     typedef unsigned int frequency;
     typedef unsigned int digital_measurement; // NOTE: this is a digital sample
     typedef unsigned int analog_measurement; // NOTE: this is a analog sample
